chore: release v0.0.5
## Summary

This release changes the behavior of additional scopes that can be defined in `.mono.toml`, allowing singles files to be scopes (e.g. `Dockerfile`). From now on, it's required to add a `**` path suffix to each scope that maps to a folder. Consider the following examples:

``` toml
[changeset.scopes]
python = "python/**"
docker = "Dockerfile"
```

Signed-off-by: squidfunk <martin.donath@squidfunk.com>